+++ CHANGELOG.md
@@ -3,12 +3,16 @@ Please see https://github.com/elelabdev/net-commander/releases for the latest re
 All notable changes to the "net-commander" extension will be documented in this file.
 
 <br><br>
-## 2025-06-20 - [0.0.3]
+## 2025-09-16 - [0.0.9]
 ### Added
+- Added a new module to query Microsoft Azure CIDRs. You can now query right from Visual Studio Code all Microsoft Azure subscriptions and search for specific CIDRs in use.
+
+## 2025-06-20 - [0.0.3]
+### Minor fix
 - Minor fixes for Root Cause Analysis for Microsoft Azure
 
 ## 2025-05-16 - [0.0.2]
-### Added
+### Minor fix
 - Minor fixes for Windows Tracert Command
 
 ## 2025-05-16 - [0.0.1]